+++ src/ambition/Ambition_vga.cc
@@ -1042,6 +1042,19 @@ void drawButtonOverlay(
   );
 }
 
+void drawFeedbackLink(
+) {
+  UserInterface::printParagraph(
+    font_news,
+    std::string("_") + _("Feedback") + "_",
+    UserInterface::FEEDBACK_LINK,
+    2,
+    UserInterface::Clear::None,
+    UserInterface::HorizontalAlignment::Left,
+    UserInterface::VerticalAlignment::Bottom
+  );
+}
+
 void drawFirmBuilderIcon(
   Firm* firm
 ) {